RAG-powered PDF Search Engine/requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
python-dotenv
requests
